requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
faiss-cpu==1.8.0
sentence-transformers==3.0.0
pydantic==2.7.1
python-dotenv==1.0.1
numpy==1.26.4
google-genai>=1.0.0
